Refactor CMake and UI components for AnalyzerPro
- Removed unnecessary PUBLIC include directive from CMakeLists.txt for mdsp_ui.
- Updated MainView to enhance control binding and improve comments for clarity.
- Refined layout management in HeaderBar and ControlRail, integrating new combo boxes for mode, FFT size, and averaging.
- Added debug auditing for APVTS parameters in MainView to ensure UI bindings are complete.
- Cleaned up unused code and improved overall responsiveness of UI components.

+++ Source/ui/meters/MeterGroupComponent.cpp
@@ -124,9 +124,9 @@ void MeterGroupComponent::resized()
     labelArea_ = headerArea_.removeFromTop (16);  // TODO: Move to metrics
     toggleArea_ = headerArea_;
 
-    metersArea_ = b.reduced (m.strokeThick, m.strokeThick);
+    metersArea_ = b.reduced (static_cast<int> (m.strokeThick), static_cast<int> (m.strokeThick));
 
-    auto toggle = toggleArea_.reduced (m.padSmall, m.strokeThick);
+    auto toggle = toggleArea_.reduced (m.padSmall, static_cast<int> (m.strokeThick));
     const int toggleW = toggle.getWidth();
     const int half = toggleW / 2;
     rmsButton_.setBounds (toggle.removeFromLeft (half));
